SOLID complete
S -> Singlton Responsibility Principle
the class need only one reason to change. like class having only one responsibility if it changes there is only one reason
When Class have multiple responsibility so it hard to understand hard to change
hard to understand
more prone bug code
ex in this we have Some classes like GameEngine, RuleEngine, AIPlayer where every class have its own resposibility the Game engine only start the game and in the ruleEngine it mainly responsible for winCheck, forkCheck etc.
O -> Open Closed principle
The class should be open for extension but close for modification
You should be add new feature without changing existing code because it does not need to test everything
Ex. Like here we define rule in the TicTacToe Board so if any new Game come the rules are define and we need not to change the RuleEngine we are taking all this rule from the board itself so not need to change the existing code
L -> Liskov Subsitution Principle
The child class must follow the behaviour of the parent class it should not be voilet the principle its parent like if the parent is move and if the child are differnt types of move then the CHild should be return valid move the cell must be empty or which is related to the parent
I -> Interface segregation Principle
doing correct interface segregation we are making code more readable and simplifild here we add one more interface CellBoard having the method getSymbol because of this our lot of generic code get simplified and now we are remove the generic which make code simple and readable. the Board is for all the board and CellBoard is for board having cell
Interface should be small, focused and specific.
Not to create fat interfaces
D -> Dependency Inversion Principle
Abstraction should not be depends on the details
Details Depends on abstraction
Make it loosing couple